Add automated plant command and refactor HouseSignEventHandler
